Move manifest_util.py into utils/
*OSBS-8605

Also change files which import manifest_util.py
