+++ atomic_reactor/utils/manifest.py
